# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_your-org_radius-go_internal_domain.CreateVoucherPackageRequest:
    properties:
      data_cap_bytes:
        type: integer
      description:
        type: string
      max_concurrent_users:
        type: integer
      name:
        type: string
      price:
        type: number
      speed_download_kbps:
        type: integer
      speed_upload_kbps:
        type: integer
      time_limit_seconds:
        type: integer
      time_limit_type:
        type: string
    type: object
...
  github_com_your-org_radius-go_internal_domain.GenerateVoucherRequest:
    properties:
      code_format:
        description: '"random" or "custom"'
        type: string
      code_length:
        description: for random
        type: integer
      count:
        type: integer
      custom_code:
        description: for custom
        type: string
      custom_password:
        description: for custom
        type: string
      package_id:
        type: string
      password_mode:
        description: '"same_as_user", "random", "custom"'
        type: string
    type: object
...
  github_com_your-org_radius-go_internal_domain.ServiceType:
    enum:
    - framed
    - login
    type: string
    x-enum-varnames:
    - ServiceTypeFramed
    - ServiceTypeLogin
...
  github_com_your-org_radius-go_internal_domain.Subscriber:
    properties:
      bandwidth_max_down:
        type: integer
      bandwidth_max_up:
        type: integer
      created_at:
        type: string
      email:
        type: string
      enabled:
        type: boolean
      expires_at:
        type: string
      framed_ip:
        type: string
      full_name:
        type: string
      id:
        type: string
      idle_timeout:
        type: integer
      is_voucher:
        type: boolean
      max_total_octets:
        type: integer
      mikrotik_group:
        type: string
      rate_limit:
        type: string
      service_type:
        $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.ServiceType'
      session_timeout:
        type: integer
      simultaneous_use:
        type: integer
      updated_at:
        type: string
      username:
        type: string
      voucher_package_id:
        type: string
    type: object
  github_com_your-org_radius-go_internal_domain.TimeLimitType:
    enum:
    - calendar
    - usage
    type: string
    x-enum-varnames:
    - TimeLimitCalendar
    - TimeLimitUsage
...
  github_com_your-org_radius-go_internal_domain.UpdateVoucherPackageRequest:
    properties:
      data_cap_bytes:
        type: integer
      description:
        type: string
      enabled:
        type: boolean
      max_concurrent_users:
        type: integer
      name:
        type: string
      price:
        type: number
      speed_download_kbps:
        type: integer
      speed_upload_kbps:
        type: integer
      time_limit_seconds:
        type: integer
      time_limit_type:
        type: string
    type: object
  github_com_your-org_radius-go_internal_domain.VoucherBalance:
    properties:
      data_bytes_remaining:
        type: integer
      data_bytes_used:
        type: integer
      data_cap_bytes:
        type: integer
      enabled:
        type: boolean
      expires_at:
        type: string
      first_login_at:
        type: string
      package_name:
        type: string
      time_limit_seconds:
        type: integer
      time_limit_type:
        type: string
      usage_seconds_remaining:
        type: integer
      usage_seconds_used:
        type: integer
      username:
        type: string
    type: object
  github_com_your-org_radius-go_internal_domain.VoucherPackage:
    properties:
      created_at:
        type: string
      data_cap_bytes:
        type: integer
      description:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      max_concurrent_users:
        type: integer
      name:
        type: string
      price:
        type: number
      speed_download_kbps:
        type: integer
      speed_upload_kbps:
        type: integer
      time_limit_seconds:
        type: integer
      time_limit_type:
        $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.TimeLimitType'
      updated_at:
        type: string
    type: object
  github_com_your-org_radius-go_internal_runtime.Envelope:
    properties:
      data: {}
      error: {}
      meta: {}
      success:
        type: boolean
    type: object
...
  internal_radius.GeneratedVoucher:
    properties:
      bandwidth_max_down:
        type: integer
      bandwidth_max_up:
        type: integer
      created_at:
        type: string
      email:
        type: string
      enabled:
        type: boolean
      expires_at:
        type: string
      framed_ip:
        type: string
      full_name:
        type: string
      id:
        type: string
      idle_timeout:
        type: integer
      is_voucher:
        type: boolean
      max_total_octets:
        type: integer
      mikrotik_group:
        type: string
      password:
        type: string
      rate_limit:
        type: string
      service_type:
        $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.ServiceType'
      session_timeout:
        type: integer
      simultaneous_use:
        type: integer
      updated_at:
        type: string
      username:
        type: string
      voucher_package_id:
        type: string
    type: object
...
host: localhost:8083
info:
  contact: {}
  description: |-
    RADIUS server with HTTP management API.
    Management endpoints require the X-Internal-Secret header.
  title: radius-go API
  version: "1.0"
paths:
...
  /voucher-packages:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.VoucherPackage'
                  type: array
              type: object
      security:
      - InternalSecret: []
      summary: List voucher packages
      tags:
      - VoucherPackages
    post:
      consumes:
      - application/json
      parameters:
      - description: Package fields
        in: body
        name: pkg
        required: true
        schema:
          $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.CreateVoucherPackageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.VoucherPackage'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
      security:
      - InternalSecret: []
      summary: Create voucher package
      tags:
      - VoucherPackages
  /voucher-packages/{id}:
    delete:
      parameters:
      - description: Package UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
      security:
      - InternalSecret: []
      summary: Delete voucher package
      tags:
      - VoucherPackages
    put:
      consumes:
      - application/json
      parameters:
      - description: Package UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated fields
        in: body
        name: pkg
        required: true
        schema:
          $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.UpdateVoucherPackageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.VoucherPackage'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
      security:
      - InternalSecret: []
      summary: Update voucher package
      tags:
      - VoucherPackages
  /vouchers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.Subscriber'
                  type: array
              type: object
      security:
      - InternalSecret: []
      summary: List vouchers
      tags:
      - Vouchers
  /vouchers/{code}/balance:
    get:
      parameters:
      - description: Voucher code (username)
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.VoucherBalance'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
      security:
      - InternalSecret: []
      summary: Get voucher remaining balance
      tags:
      - Vouchers
  /vouchers/generate:
    post:
      consumes:
      - application/json
      parameters:
      - description: Generation request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.GenerateVoucherRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_radius.GeneratedVoucher'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
      security:
      - InternalSecret: []
      summary: Generate voucher(s)
      tags:
      - Vouchers
schemes:
- http
securityDefinitions:
  InternalSecret:
    in: header
    name: X-Internal-Secret
    type: apiKey
swagger: "2.0"